Move images CLI parser to submodule
This will avoid circular imports in the next steps of the refactor.
