+++ setup.cfg
@@ -71,5 +71,5 @@ console_scripts =
     check-tristan-files = tristan.diagnostics.check_files:cli
     cues = tristan.command_line.cues:main
     find-trigger-intervals = tristan.diagnostics.find_trigger_intervals:cli
-    images = tristan.command_line.images:main
+    images = tristan.command_line.images.top_level_cli:main
     valid-events = tristan.diagnostics.check_valid_events:cli
